backend/requirements.txt
fastapi
<<<<<<< HEAD
uvicorn[standard]
gunicorn
resend
supabase
pymongo
dnspython
python-dotenv
beautifulsoup4
requests
pydantic
email-validator
itsdangerous
apscheduler
pywebpush
stripe
bleach
slowapi
langdetect
google-generativeai
user-agents
sentry-sdk[fastapi]
certifi
psycopg[binary]
=======
uvicorn
pydantic
pydantic-settings
supabase
pyjwt
python-dotenv
sqlmodel
asyncpg
sqlalchemy[asyncio]
bleach
stripe
requests
beautifulsoup4
langdetect
ftfy
unstructured
gunicorn
httpx
>>>>>>> 4c20d82 (Jobshaman MVP 2.0: Clean repo, i18n Nordic expansion & engine optimization)
